Insert blocks into DB only _once_ processed by OrderWatcher so that when Mesh exits and spins back up, it'll re-process blocks that weren't processed
